Improve figures (#20)
* change MCMC parameters

* fixed notebook

* changed parameters

* update

* minor fixes

* round automatically casts to int

* rewroted clustered network algo

* rewrote clustered network heatmap

* fix bugs in clustered network algo

* added prettify_matrix function

* rename p

* updates

* updates

* update data

* updates

* update figures

* added new data

* redo figures

* update figures

* update

* updated figs

* rename file

* updates

+++ lcs/inference.py
@@ -537,7 +537,7 @@ def infer_dyamics_loop(x, A, p_gamma, p_c):
 
         # infection events
         for i, nu in enumerate(nus):
-            nu = int(round(nu))
+            nu = round(nu)
             a[nu] += x[t + 1, i] * (1 - x[t, i])
             b[nu] += (1 - x[t + 1, i]) * (1 - x[t, i])
